Working on mmix-port (currently on paging)

+++ source/kernel/src/printf.c
@@ -151,7 +151,7 @@ void prf_vprintf(sPrintEnv *env,const char *fmt,va_list ap) {
 					fmt++;
 					break;
 				case '*':
-					pad = va_arg(ap, uint);
+					pad = va_arg(ap, ulong);
 					fmt++;
 					break;
 				case '|':
